flask-rag/requirements.txt
langchain-core>=0.1.27
langchain-ollama>=0.1.3
langchain>=0.1.9
langchain-community>=0.0.24
chromadb>=0.4.24
flask>=3.0.2
flask-cors>=4.0.0
PyPDF>=3.0.1
google-cloud-storage>=2.15.0